requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
tiktoken>=0.5.0
PyJWT>=2.6.0
tqdm>=4.65.0
requests>=2.28.0
beautifulsoup4>=4.12.0
langchain-ollama>=0.0.1
python-dateutil>=2.8.2
celery>=5.3.0
redis>=4.5.0